Home

ReferenceTests.jl Documentation

ReferenceTests.jl is a Julia package that adds a couple of additional macros to your testing toolbox. In particular, it focuses on functionality for testing values against reference files, which in turn the package can help create and update if need be. ReferenceTests.jl is build on top of FileIO.jl and designed to be used alongside Base.Test.

Contents

ReferenceTests.jl Documentation

Contents

Index

Public Interface

Index

ReferenceTests.@io2str

ReferenceTests.@test_reference

ReferenceTests.@withcolor

Public Interface

ReferenceTests.@withcolor — Macro.

@withcolor ex

Make sure that ex is evaluated while Base.have_color is set to true. The original value of Base.have_color will be restored afterwards.

This macro is particularily useful for CI, where it is not unusual that julia is executed without the --color=yes argument by default.

@withcolor print_with_color(:green, "foo")

source

ReferenceTests.@io2str — Macro.

@io2str ex

Search and replace the placeholder ::IO with a newly allocated Base.IOBuffer which is afterwards converted to a string and returned.

This macro allows a user to conveniently test text-based printing functionality that require an IO argument. A particularly common example of such would be a custom Base.show method.

Examples

julia> using ReferenceTests

julia> @io2str print(::IO, "Hello World")
"Hello World"

julia> @io2str show(IOContext(::IO, :limit=>true, :displaysize=>(10,10)), "text/plain", ones(10))
"10-element Array{Float64,1}:\n 1.0\n 1.0\n 1.0\n ⋮  \n 1.0\n 1.0"

source

ReferenceTests.@test_reference — Macro.

@test_reference filename expr [by] [kw...]

Tests that the expression expr with reference filename using equality test strategy by.

The pipeline of test_reference is:

preprocess expr

read and preprocess filename

compare the results using by

if test fails in an interactive session (e.g, include(test/runtests.jl)), an interactive dialog will be trigered.

Arguments:

filename::String: relative path to the file that contains the macro invocation.

expr: the actual content used to compare.

by: the equality test function. By default it is isequal if not explicitly stated.

Types

The file-extension of filename, as well as the type of the result of evaluating expr, determine how contents are processed and compared. The contents is treated as:

Images when expr is an image type, i.e., AbstractArray{<:Colorant};

SHA256 when filename endswith *.sha256;

Text as a fallback.

Images

Images are compared approximately using a different by to ignore most encoding and decoding errors. The default value is Images.@test_approx_eq_sigma_eps.

The file can be either common image files (e.g., *.png), which are handled by FileIO, or text-coded *.txt files, which is handled by ImageInTerminal. Text-coded image has less storage requirements and also allows to view the reference file in a simple terminal using cat.

SHA256

The hash of the expr and content of filename are compared.

Tip

This is useful for a convenient low-storage way of making sure that the return value doesn't change for selected test cases.

Fallback

Simply test the equality of expr and the contents of filename without any preprocessing.

Examples

# store as string using ImageInTerminal with encoding size (5,10)
@test_reference "camera.txt" testimage("cameraman") size=(5,10)

# using folders in the relative path is allowed
@test_reference "references/camera.png" testimage("cameraman")

# Images can also be stored as hash. Note however that this
# can only check for equality (no tolerance possible)
@test_reference "references/camera.sha256" testimage("cameraman")

# test images using ImageQualityIndexes.PSNR
@test_reference "references/camera.png" testimage("cameraman") by=(x,y)->psnr(x,y)>25

# test number with absolute tolerance 10
@test_reference "references/string3.txt" 1338 by=(ref, x)->isapprox(ref, x; atol=10)

source