Home

ReferenceTests.jl Documentation

ReferenceTests.jl is a Julia package that adds a couple of additional macros to your testing toolbox. In particular, it focuses on functionality for testing values against reference files, which in turn the package can help create and update if need be. ReferenceTests.jl is build on top of FileIO.jl and designed to be used alongside Base.Test.

Contents

ReferenceTests.jl Documentation

Contents

Index

Public Interface

Index

ReferenceTests.@io2str

ReferenceTests.@test_reference

ReferenceTests.@withcolor

Public Interface

ReferenceTests.@withcolor — Macro.

@withcolor ex

Make sure that ex is evaluated while Base.have_color is set to true. The original value of Base.have_color will be restored afterwards.

This macro is particularily useful for CI, where it is not unusual that julia is executed without the --color=yes argument by default.

@withcolor print_with_color(:green, "foo")

source

ReferenceTests.@io2str — Macro.

@io2str ex

Search and replace the placeholder ::IO with a newly allocated Base.IOBuffer which is afterwards converted to a string and returned.

This macro allows a user to conveniently test text-based printing functionality that require an IO argument. A particularly common example of such would be a custom Base.show method.

Examples

julia> using ReferenceTests

julia> @io2str print(::IO, "Hello World")
"Hello World"

julia> @io2str show(IOContext(::IO, :limit=>true, :displaysize=>(10,10)), "text/plain", ones(10))
"10-element Array{Float64,1}:\n 1.0\n 1.0\n 1.0\n ⋮  \n 1.0\n 1.0"

source

ReferenceTests.@test_reference — Macro.

@test_reference filename ex [kw...]

Tests that the expression ex evaluates to the same result as stored in the given reference file, which is denoted by the string filename. Executing the code in an interactive julia session will trigger an interactive dialog if results don't match. This dialog allows the user to create and/or update the reference files.

The given string filename is assumed to be the relative path to the file that contains the macro invocation. This likely means that the path filename is relative to the test/ folder of your package.

The file-extension of filename, as well as the type of the result of evaluating ex, determine how the actual value is compared to the reference value. The default implementation will do a simple equality check with the result of FileIO.load. This means that it is the user's responsibility to have the required IO package installed.

Colorant arrays (i.e.) receive special treatment. If the extension of filename is txt then the package ImageInTerminal will be used to create a string based cure approximation of the image. This will have low storage requirements and also allows to view the reference file in a simple terminal using cat.

Another special file extension is sha256 which will cause the hash of the result of ex to be stored and compared as plain text. This is useful for a convenient low-storage way of making sure that the return value doesn't change for selected test cases.

Examples

# store as string using ImageInTerminal with encoding size (5,10)
@test_reference "camera.txt" testimage("cameraman") size=(5,10)

# using folders in the relative path is allowed
@test_reference "references/camera.png" testimage("cameraman")

# Images can also be stored as hash. Note however that this
# can only check for equality (no tolerance possible)
@test_reference "references/camera.sha256" testimage("cameraman")

source